Refactor auth module: rename file, clean up imports and comments
