+++ apps/users-service/src/routes/auth.ts
